Instance specific properties container / for integrating other plugins
