+++ src/test/resources/com/cognifide/gradle/aem/test/debug/additional/gradle.properties
@@ -6,4 +6,6 @@ aem.instance.local.local-publish1.httpUrl=http://localhost:4503
 aem.instance.local.local-publish2.httpUrl=http://localhost:4504
 
 aem.instance.remote.test-author.httpUrl=http://192.168.6.6:4502
+aem.instance.remote.test-author.externalUrl=http://test-author.local
 aem.instance.remote.test-publish.httpUrl=http://192.168.6.6:4503
+aem.instance.remote.test-publish.externalUrl=http://test-publish.local
